chore(docs): Update tutorial documentation part 4, 7 & 8 (#14257)
* Amended code highlights for staticQuery section

I noticed running through the tutorial that lines 306 and 353 do actually change, but were not highlighted in the code block.

The beginning of the export has a curly brace at the front at this point, as does the end, but earlier in the tutorial it opens and closes with parentheses. The code worked when copying and pasting but I was typing in the highlighted changes, in an attempt to learn through doing, so I'm updating the documentation, just in case anyone is doing the same and would like it to be a little clearer!

I've also removed a redundant //highlight-start and //highlight-end at 318 and 319

* Added note to restart the development server

I got caught on this for a few minutes, as it doesn't explicitly say you need to restart the development server to verify whether this worked or not. Added a note.

* Made localhost a link and clearer tutorial code blocks

As I was reading this, it felt like the sections in the tutorial were less linked to each other than in the previous sections, given the three packages mentioned all related to the lighthouse auditing tool. So I made the code-blocks include the same continual code, with highlighting to add what changed in each block, rather than the blocks in isolated.

I also made the localhost:9000 a proper link, and added a dummy Gatsby coloured icon to the tutorial so people could test this out better without having to source one.

* localhost:9000 as url

Last commit wasn't formatted correctly.

* adding dummy icon file

* Tweaked offline-plugin note relating to order

As this section refers to the offline plugin, it feels like it makes sense to denote where it goes, rather than what comes before it.

* Updating icon image path

* Updating icon img path

* Wrapping localhost link in ``

Co-Authored-By: Lennart <lekoarts@gmail.com>

* Updating link text to 'this example icon'

Sounds a clearer than just 'this'

Co-Authored-By: Lennart <lekoarts@gmail.com>

* updating icon to gatsbyjs icon

* updating code comment to respect indentation

* Updating some config values from ` to '

In the `gatsby-config` file there were some needless ` surrounding object values. I've amended these to ' according to the Prettier formatting playground.

* chore: format

* swapped " for ` (back-ticks) in gatsby-config.js

In gastby-config.js, swapped " for ` (back-ticks) to make consistent with other references in the rest of the tutorial docs.

* remove errant "

Removed a couple of errant " within backticks.